[REFACTOR] config, filter 클래스 패키지 변경
