+++ src/main/java/kr/hhplus/be/ecommerce/config/RedisRedissonConfig.java
@@ -8,7 +8,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class RedissonConfig {
+public class RedisRedissonConfig {
 
     @Value("${spring.data.redis.host}")
     private String redisHost;
